extract components from Tabs
